requirements.txt
fastapi
uvicorn
google-cloud-datastore
six
python-jose